pyproject.toml
[tool.poetry]
name = "1001-albums-slack-bot"
version = "0.1.0"
description = ""
authors = ["Ren Kasiyas <okasiyas@gmail.com>"]
readme = "README.md"
packages = [{include = "1001_albums_slack_bot"}]

[tool.poetry.dependencies]
python = "^3.9"
slack-sdk = "^3.19.5"
python-dotenv = "^0.21.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
